predict.py
from sklearn.preprocessing import LabelEncoder, StandardScaler
import joblib
import pandas as pd

# Placeholder function for predicting rainfall (replace with your model)
def predict_rainfall(features):
    """
    Predicts the amount of rainfall based on the given features.

    Args:
        features (pandas.DataFrame): A DataFrame containing the features used for prediction.

    Returns:
        numpy.ndarray: An array of predicted rainfall values.
    """
    encoder = LabelEncoder()
    features['Events'] = encoder.fit_transform(features['Events'])

    sc = StandardScaler()
    x_std = sc.fit_transform(features)
    x_std = pd.DataFrame(x_std, columns=features.columns)

    model = joblib.load('model_new.pkl')

    rainfall = model.predict(x_std)
    return rainfall